SimpleLottery: Renaming wrapper as Lottery
